requirements.txt
google-genai
gradio
rfdetr
supervision
gdown